Support for custom Globally Unique ID field name
According to https://relay.dev/docs/guides/graphql-server-specification/#object-identification
the Node interface should have a field named `id` with the Globally
Unique ID. However, for simpler APIs it can be beneficial to have the
name of this field customized to expose the raw object id under the
field `id`, and the global id under a different name (e.g. `global_id`).

This is achieved by using a metaclass for the Node abstract class which
defines the field according to the value of
`config.RELAY_GLOBAL_ID_FIELD_NAME`.